+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
 ## 3.3.0
 ### Added
 - Refactor of `bin/aws-firewall-factory.ts`, grouping duplicated code on a function, adding comments and better organizing the file.
@@ -13,6 +14,23 @@
 
 =======
 =======
+=======
+## 3.2.3
+### Fixed
+- Bump eslint from 8.41.0 to 8.42.0
+- Bump @typescript-eslint/eslint-plugin from 4.4.0 to 5.59.11
+- Bump @typescript-eslint/parser from 4.4.0 to 5.59.11
+- Bump eslint-plugin-import from 2.24.2 to 2.27.5
+- Bump jest from 26.4.2 to 29.5.0
+- Bump ts-jest from 26.2.0 to 29.1.0
+- pinned dependencies to avoid conflicting peer dependencies
+- fix eslint issues
+- fix logging bucket encryption issues referencing [23513](https://github.com/aws/aws-cdk/issues/23513)
+- fix cloudformation warning when using statements that inspect single headers (lowercase name key)
+
+### Added
+- Added Linting command `lint` to npm scripts which can be run via `npm run lint`◊
+>>>>>>> 2d597da6 (fix: dependencies, single-header cloudformation warning and log bucket encryption bug)
 ## 3.2.2
 ### Fixed
 - Bump @aws-sdk/client-pricing from 3.332.0 to 3.341.0